# #411 use safer method when converting between byte and float slices