--- a/.ci/geos.Dockerfile
+++ b/.ci/geos.Dockerfile
@@ -1,4 +1,4 @@
-FROM golang:1.14
+FROM golang:1.17
 RUN apt-get -y update && \
 	apt-get install -y libgeos-dev=3.7.1-1 && \
 	rm -rf /var/lib/apt/lists/*