--- a/geom/wkb_marshal.go
+++ b/geom/wkb_marshal.go
@@ -4,7 +4,6 @@ import (
 	"encoding/binary"
 	"fmt"
 	"math"
-	"reflect"
 	"unsafe"
 )
 
@@ -87,10 +86,5 @@ func (m *wkbMarshaler) writeSequence(seq Sequence) {
 // floatsAsBytes reinterprets the floats slice as a bytes slice in a similar
 // manner to reinterpret_cast in C++.
 func floatsAsBytes(floats []float64) []byte {
-	var byts []byte
-	hdr := (*reflect.SliceHeader)(unsafe.Pointer(&byts))
-	hdr.Data = (*reflect.SliceHeader)(unsafe.Pointer(&floats)).Data
-	hdr.Len = 8 * len(floats)
-	hdr.Cap = 8 * cap(floats)
-	return byts
+	return unsafe.Slice((*byte)(unsafe.Pointer(&floats[0])), 8*len(floats))
 }