--- a/go.mod
+++ b/go.mod
@@ -1,5 +1,5 @@
 module github.com/peterstace/simplefeatures
 
-go 1.14
+go 1.17
 
 require github.com/lib/pq v1.1.1