## Description

We want to move away from editing slice headers when converting between byte and float slices. And to use the "safer"  unsafe.slice method available in golang 1.17.

Benchmark comparisons current approach vs the new one
```
Benchmark_FloatAsBytes_SliceHeaders-12                   1000000000               0.3534 ns/op
Benchmark_FloatAsBytes_UnsafeSlice-12                    437308378                2.538 ns/op
Benchmark_FloatAsBytesUsingStandardLibrary-12            2905132                  395.0 ns/op

```
Using the new method while slower than the existing edit header approach is still considerably faster than using the standard library.

## Check List

Have you:

- Added unit tests?

- Add cmprefimpl tests? (if appropriate?)

- Updated release notes? (if appropriate?)

## Related Issue

#411 

## Benchmark Results

- Please paste benchmark results here. The benchmarks can be run using the
  `run_benchmarks.sh` script.

<details>

<summary>Click to expand</summary>

```
RISON
name                                                         old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
LineEnvelope/0-12                                              3.56ns ±25%    3.50ns ±18%      ~     (p=0.715 n=15+14)
LineEnvelope/1-12                                              3.82ns ±20%    3.49ns ±22%    -8.74%  (p=0.028 n=15+14)
LineEnvelope/2-12                                              3.71ns ±25%    3.35ns ±10%    -9.74%  (p=0.032 n=15+12)
LineEnvelope/3-12                                              3.36ns ±25%    3.38ns ±25%      ~     (p=0.949 n=15+14)
MarshalWKB/polygon/n=10-12                                      290ns ±14%     290ns ±20%      ~     (p=0.967 n=15+15)
MarshalWKB/polygon/n=100-12                                     560ns ±10%     549ns ± 7%      ~     (p=0.252 n=15+14)
MarshalWKB/polygon/n=1000-12                                   2.93µs ± 7%    2.92µs ± 7%      ~     (p=0.974 n=15+13)
MarshalWKB/polygon/n=10000-12                                  19.7µs ± 2%    20.2µs ±14%      ~     (p=0.399 n=12+15)
UnmarshalWKB/polygon/n=10-12                                    457ns ±12%     440ns ±18%      ~     (p=0.270 n=15+14)
UnmarshalWKB/polygon/n=100-12                                   756ns ± 9%     788ns ±17%      ~     (p=0.270 n=14+15)
UnmarshalWKB/polygon/n=1000-12                                 3.31µs ± 8%    3.34µs ± 7%      ~     (p=0.323 n=14+14)
UnmarshalWKB/polygon/n=10000-12                                26.8µs ±11%    25.8µs ± 4%      ~     (p=0.427 n=15+12)
IntersectsLineStringWithLineString/n=10-12                     1.97µs ±13%    1.87µs ± 3%    -5.14%  (p=0.025 n=15+12)
IntersectsLineStringWithLineString/n=100-12                    26.7µs ±17%    26.6µs ±15%      ~     (p=0.839 n=14+14)
IntersectsLineStringWithLineString/n=1000-12                    281µs ±22%     285µs ±21%      ~     (p=0.683 n=14+15)
IntersectsLineStringWithLineString/n=10000-12                  4.25ms ±13%    4.12ms ± 9%      ~     (p=0.270 n=15+14)
IntersectsMultiPointWithMultiPoint/n=20-12                     1.20µs ±27%    1.19µs ±16%      ~     (p=0.705 n=15+15)
IntersectsMultiPointWithMultiPoint/n=200-12                    13.2µs ±12%    13.2µs ± 7%      ~     (p=0.979 n=13+12)
IntersectsMultiPointWithMultiPoint/n=2000-12                    125µs ± 4%     123µs ± 8%      ~     (p=0.118 n=12+14)
IntersectsMultiPointWithMultiPoint/n=20000-12                  1.46ms ±16%    1.43ms ± 6%      ~     (p=0.616 n=14+13)
PolygonSingleRingValidation/n=10-12                            3.33µs ±14%    3.41µs ±11%      ~     (p=0.310 n=14+15)
PolygonSingleRingValidation/n=100-12                           43.9µs ± 9%    44.9µs ±15%      ~     (p=0.331 n=14+15)
PolygonSingleRingValidation/n=1000-12                           562µs ±26%     559µs ±18%      ~     (p=0.910 n=14+14)
PolygonSingleRingValidation/n=10000-12                         7.31ms ±12%    7.49ms ±15%      ~     (p=0.285 n=15+15)
PolygonMultipleRingsValidation/n=4-12                          10.4µs ±27%     9.9µs ±13%      ~     (p=0.345 n=15+15)
PolygonMultipleRingsValidation/n=36-12                         80.4µs ±22%    83.9µs ±23%      ~     (p=0.187 n=15+15)
PolygonMultipleRingsValidation/n=400-12                        1.06ms ±11%    1.07ms ±14%      ~     (p=0.914 n=14+15)
PolygonMultipleRingsValidation/n=4096-12                       12.2ms ± 6%    12.2ms ±24%      ~     (p=0.290 n=14+15)
PolygonZigZagRingsValidation/n=10-12                           15.0µs ±10%    15.8µs ±19%      ~     (p=0.098 n=13+15)
PolygonZigZagRingsValidation/n=100-12                           176µs ±10%     180µs ±18%      ~     (p=0.545 n=13+13)
PolygonZigZagRingsValidation/n=1000-12                         2.06ms ±14%    2.06ms ± 9%      ~     (p=0.949 n=15+14)
PolygonZigZagRingsValidation/n=10000-12                        26.1ms ±13%    25.4ms ± 9%      ~     (p=0.403 n=14+12)
PolygonAnnulusValidation/n=10-12                               5.10µs ±22%    5.31µs ±18%      ~     (p=0.178 n=14+14)
PolygonAnnulusValidation/n=100-12                              45.3µs ±12%    45.5µs ±24%      ~     (p=0.683 n=15+15)
PolygonAnnulusValidation/n=1000-12                              791µs ±17%     783µs ±14%      ~     (p=0.914 n=15+14)
PolygonAnnulusValidation/n=10000-12                            9.35ms ±12%    9.18ms ±10%      ~     (p=0.454 n=14+14)
MultipolygonValidation/n=1-12                                   606ns ±18%     575ns ±16%      ~     (p=0.051 n=15+14)
MultipolygonValidation/n=4-12                                  1.37µs ±23%    1.36µs ±14%      ~     (p=0.744 n=15+15)
MultipolygonValidation/n=16-12                                 5.64µs ±25%    5.26µs ±13%      ~     (p=0.252 n=15+14)
MultipolygonValidation/n=64-12                                 24.5µs ±15%    24.5µs ±12%      ~     (p=0.976 n=15+15)
MultipolygonValidation/n=256-12                                 158µs ± 9%     162µs ±15%      ~     (p=0.325 n=15+15)
MultipolygonValidation/n=1024-12                                716µs ±14%     719µs ±12%      ~     (p=1.000 n=15+15)
MultiPolygonTwoCircles/n=10-12                                 4.97µs ±26%    5.08µs ±10%      ~     (p=0.964 n=15+13)
MultiPolygonTwoCircles/n=100-12                                52.0µs ±19%    54.0µs ±17%      ~     (p=0.233 n=15+15)
MultiPolygonTwoCircles/n=1000-12                                551µs ±15%     565µs ±13%      ~     (p=0.254 n=13+15)
MultiPolygonTwoCircles/n=10000-12                              7.82ms ±16%    8.29ms ±10%    +6.00%  (p=0.026 n=15+14)
MultiPolygonMultipleTouchingPoints/n=1-12                      6.94µs ±22%    6.84µs ±18%      ~     (p=0.806 n=15+15)
MultiPolygonMultipleTouchingPoints/n=10-12                     53.3µs ±20%    53.0µs ±15%      ~     (p=0.967 n=15+15)
MultiPolygonMultipleTouchingPoints/n=100-12                     625µs ±17%     643µs ±23%      ~     (p=0.486 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1000-12                   7.54ms ±14%    7.53ms ±13%      ~     (p=0.870 n=15+15)
WKTParsing/point-12                                            2.31µs ± 6%    2.35µs ±17%      ~     (p=0.563 n=11+14)
DistancePolygonToPolygonOrdering/n=100_swap=false-12           67.4µs ±15%    68.6µs ±15%      ~     (p=0.713 n=15+15)
DistancePolygonToPolygonOrdering/n=100_swap=true-12            65.7µs ±19%    68.1µs ±23%      ~     (p=0.539 n=15+15)
DistancePolygonToPolygonOrdering/n=1000_swap=false-12          1.12ms ±21%    1.05ms ±10%      ~     (p=0.062 n=14+14)
DistancePolygonToPolygonOrdering/n=1000_swap=true-12           1.03ms ±15%    1.05ms ±17%      ~     (p=0.525 n=13+15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false-12     7.12µs ±23%    6.81µs ± 7%      ~     (p=0.234 n=14+15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true-12      7.33µs ±25%    6.81µs ±19%      ~     (p=0.069 n=15+15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false-12    78.3µs ±11%    75.5µs ±20%      ~     (p=0.146 n=15+14)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true-12     76.2µs ±25%    75.7µs ±12%      ~     (p=1.000 n=15+15)
MultiLineStringIsSimpleManyLineStrings/n=100-12                75.3µs ±20%    69.4µs ± 9%    -7.86%  (p=0.012 n=14+13)
MultiLineStringIsSimpleManyLineStrings/n=1000-12                765µs ±20%     770µs ±15%      ~     (p=0.687 n=13+13)
ForceCWandForceCCW/0-12                                        45.8ns ±21%    44.9ns ±21%      ~     (p=0.488 n=14+13)
ForceCWandForceCCW/0#01-12                                      317ns ±19%     325ns ±19%      ~     (p=0.760 n=14+14)
ForceCWandForceCCW/1-12                                        46.3ns ±22%    45.2ns ±13%      ~     (p=0.631 n=14+12)
ForceCWandForceCCW/1#01-12                                      318ns ±13%     310ns ±16%      ~     (p=0.185 n=13+14)
ForceCWandForceCCW/2-12                                        86.2ns ±25%    87.6ns ±33%      ~     (p=0.981 n=13+14)
ForceCWandForceCCW/2#01-12                                      496ns ±14%     516ns ±15%      ~     (p=0.347 n=12+14)
ForceCWandForceCCW/3-12                                        95.4ns ±26%    88.2ns ±22%      ~     (p=0.186 n=13+13)
ForceCWandForceCCW/3#01-12                                      536ns ±19%     508ns ±15%      ~     (p=0.133 n=12+13)
ForceCWandForceCCW/4-12                                        350ns ±132%     129ns ±34%   -63.10%  (p=0.049 n=15+15)
ForceCWandForceCCW/4#01-12                                      612ns ±82%     843ns ±25%      ~     (p=0.161 n=15+15)
ForceCWandForceCCW/5-12                                        422ns ±149%     121ns ± 6%   -71.26%  (p=0.001 n=15+12)
ForceCWandForceCCW/5#01-12                                      572ns ±80%     769ns ±15%      ~     (p=0.895 n=15+12)
ForceCWandForceCCW/6-12                                         176ns ±11%     184ns ±32%      ~     (p=0.667 n=12+14)
ForceCWandForceCCW/6#01-12                                     1.33µs ±15%    1.25µs ±17%      ~     (p=0.151 n=13+14)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
IntersectionWithoutValidation/n=10-12                          54.9µs ± 7%    54.4µs ±13%      ~     (p=0.586 n=13+15)
IntersectionWithoutValidation/n=100-12                          109µs ±21%     109µs ±19%      ~     (p=0.621 n=14+15)
IntersectionWithoutValidation/n=1000-12                         485µs ±23%     466µs ±18%      ~     (p=0.486 n=15+15)
IntersectionWithoutValidation/n=10000-12                       4.18ms ±17%    4.17ms ±16%      ~     (p=0.621 n=15+14)
NoOp/n=10-12                                                   6.86µs ±23%    6.39µs ±17%      ~     (p=0.067 n=15+15)
NoOp/n=100-12                                                  18.3µs ±16%    17.8µs ±15%      ~     (p=0.345 n=15+15)
NoOp/n=1000-12                                                  123µs ±23%     118µs ±20%      ~     (p=0.202 n=15+15)
NoOp/n=10000-12                                                1.34ms ±10%    1.38ms ±14%      ~     (p=0.316 n=13+15)
pkg:github.com/peterstace/simplefeatures/internal/perf goos:linux goarch:amd64
LineStringIsSimpleCircle/n=10-12                               2.88µs ±20%    2.85µs ±25%      ~     (p=0.645 n=15+15)
LineStringIsSimpleCircle/n=100-12                              41.3µs ±20%    38.9µs ± 5%      ~     (p=0.200 n=15+12)
LineStringIsSimpleCircle/n=1000-12                              526µs ±22%     516µs ±25%      ~     (p=0.412 n=15+15)
LineStringIsSimpleCircle/n=10000-12                            7.18ms ±15%    7.17ms ±16%      ~     (p=1.000 n=15+15)
LineStringIsSimpleZigZag/10-12                                 2.57µs ±15%    2.49µs ± 7%      ~     (p=0.168 n=12+13)
LineStringIsSimpleZigZag/100-12                                43.9µs ±26%    42.2µs ±20%      ~     (p=0.389 n=15+15)
LineStringIsSimpleZigZag/1000-12                                519µs ±29%     487µs ±17%      ~     (p=0.285 n=15+15)
LineStringIsSimpleZigZag/10000-12                              6.92ms ±16%    6.57ms ±17%      ~     (p=0.150 n=14+14)
SetOperation/n=4/Go_Intersection-12                            56.5µs ±19%    53.8µs ±22%      ~     (p=0.161 n=15+15)
SetOperation/n=4/Go_Difference-12                              59.1µs ±13%    57.0µs ±26%      ~     (p=0.116 n=13+14)
SetOperation/n=4/Go_SymmetricDifference-12                     75.9µs ±14%    74.7µs ±14%      ~     (p=1.000 n=14+13)
SetOperation/n=4/Go_Union-12                                   59.5µs ±16%    60.2µs ±23%      ~     (p=0.847 n=14+15)
SetOperation/n=4/GEOS_Intersection-12                          52.4µs ±22%    47.0µs ±28%      ~     (p=0.134 n=14+15)
SetOperation/n=4/GEOS_Difference-12                            53.6µs ±23%    51.9µs ±34%      ~     (p=0.539 n=15+15)
SetOperation/n=4/GEOS_SymmetricDifference-12                   77.5µs ± 9%    69.7µs ±25%   -10.06%  (p=0.027 n=14+14)
SetOperation/n=4/GEOS_Union-12                                 53.8µs ±14%    51.9µs ±34%      ~     (p=0.561 n=14+15)
SetOperation/n=8/Go_Intersection-12                            72.9µs ±20%    68.1µs ±10%      ~     (p=0.254 n=15+13)
SetOperation/n=8/Go_Difference-12                              73.5µs ±12%    70.0µs ±23%      ~     (p=0.467 n=13+15)
SetOperation/n=8/Go_SymmetricDifference-12                     98.4µs ±26%    91.7µs ±24%      ~     (p=0.116 n=15+15)
SetOperation/n=8/Go_Union-12                                   76.6µs ±22%    71.8µs ±24%      ~     (p=0.250 n=15+15)
SetOperation/n=8/GEOS_Intersection-12                          60.2µs ±22%    58.9µs ±22%      ~     (p=0.701 n=14+14)
SetOperation/n=8/GEOS_Difference-12                            63.6µs ±37%    58.0µs ±34%      ~     (p=0.250 n=15+15)
SetOperation/n=8/GEOS_SymmetricDifference-12                   86.3µs ±24%    82.3µs ±27%      ~     (p=0.345 n=15+15)
SetOperation/n=8/GEOS_Union-12                                 60.3µs ±32%    55.5µs ±48%      ~     (p=0.217 n=15+15)
SetOperation/n=16/Go_Intersection-12                           94.4µs ±31%    89.9µs ±27%      ~     (p=0.285 n=15+15)
SetOperation/n=16/Go_Difference-12                              104µs ±40%      99µs ±19%      ~     (p=0.653 n=15+15)
SetOperation/n=16/Go_SymmetricDifference-12                     137µs ±29%     143µs ±28%      ~     (p=0.400 n=15+14)
SetOperation/n=16/Go_Union-12                                   105µs ±21%     101µs ±20%      ~     (p=0.571 n=14+14)
SetOperation/n=16/GEOS_Intersection-12                         62.5µs ±40%    63.5µs ±43%      ~     (p=0.567 n=15+15)
SetOperation/n=16/GEOS_Difference-12                           64.0µs ±45%    64.1µs ±33%      ~     (p=0.967 n=15+15)
SetOperation/n=16/GEOS_SymmetricDifference-12                  92.5µs ±14%   103.5µs ±23%      ~     (p=0.088 n=13+15)
SetOperation/n=16/GEOS_Union-12                                60.9µs ±28%    62.6µs ±36%      ~     (p=0.511 n=14+14)
SetOperation/n=32/Go_Intersection-12                            156µs ±28%     153µs ±27%      ~     (p=0.806 n=15+15)
SetOperation/n=32/Go_Difference-12                              164µs ±30%     181µs ±24%   +10.59%  (p=0.033 n=14+15)
SetOperation/n=32/Go_SymmetricDifference-12                     221µs ±24%     233µs ±19%      ~     (p=0.217 n=15+15)
SetOperation/n=32/Go_Union-12                                   163µs ±19%     174µs ±16%      ~     (p=0.186 n=14+15)
SetOperation/n=32/GEOS_Intersection-12                         70.9µs ±36%    78.8µs ±45%      ~     (p=0.250 n=15+15)
SetOperation/n=32/GEOS_Difference-12                           71.7µs ±13%    71.7µs ±17%      ~     (p=0.943 n=14+13)
SetOperation/n=32/GEOS_SymmetricDifference-12                   131µs ±15%     132µs ±24%      ~     (p=0.813 n=15+14)
SetOperation/n=32/GEOS_Union-12                                76.3µs ±31%    77.4µs ±25%      ~     (p=0.838 n=15+15)
SetOperation/n=64/Go_Intersection-12                            256µs ±18%     250µs ±22%      ~     (p=0.325 n=15+15)
SetOperation/n=64/Go_Difference-12                              265µs ±19%     274µs ±20%      ~     (p=0.591 n=14+15)
SetOperation/n=64/Go_SymmetricDifference-12                     438µs ±36%     436µs ±39%      ~     (p=1.000 n=15+15)
SetOperation/n=64/Go_Union-12                                   277µs ±27%     288µs ±17%      ~     (p=0.246 n=14+14)
SetOperation/n=64/GEOS_Intersection-12                         87.4µs ± 9%    89.7µs ±14%      ~     (p=0.413 n=13+15)
SetOperation/n=64/GEOS_Difference-12                            111µs ±19%     105µs ± 6%    -5.69%  (p=0.029 n=15+14)
SetOperation/n=64/GEOS_SymmetricDifference-12                   226µs ±14%     215µs ± 9%      ~     (p=0.051 n=15+14)
SetOperation/n=64/GEOS_Union-12                                 118µs ±17%     115µs ± 8%      ~     (p=0.683 n=15+14)
SetOperation/n=128/Go_Intersection-12                           527µs ±39%     451µs ±18%   -14.33%  (p=0.041 n=15+14)
SetOperation/n=128/Go_Difference-12                             498µs ±28%     557µs ±36%      ~     (p=0.185 n=13+15)
SetOperation/n=128/Go_SymmetricDifference-12                    851µs ±66%     749µs ±19%      ~     (p=0.914 n=15+14)
SetOperation/n=128/Go_Union-12                                  588µs ±48%     544µs ±19%      ~     (p=0.496 n=15+13)
SetOperation/n=128/GEOS_Intersection-12                         146µs ±17%     144µs ±16%      ~     (p=0.624 n=15+15)
SetOperation/n=128/GEOS_Difference-12                           171µs ±12%     173µs ±18%      ~     (p=1.000 n=15+15)
SetOperation/n=128/GEOS_SymmetricDifference-12                  370µs ±19%     346µs ±14%      ~     (p=0.067 n=15+15)
SetOperation/n=128/GEOS_Union-12                                183µs ±22%     182µs ±15%      ~     (p=0.838 n=15+15)
SetOperation/n=256/Go_Intersection-12                           935µs ±33%     875µs ±25%      ~     (p=0.331 n=15+14)
SetOperation/n=256/Go_Difference-12                             986µs ±30%    1010µs ±37%      ~     (p=0.683 n=15+14)
SetOperation/n=256/Go_SymmetricDifference-12                   1.43ms ±24%    1.61ms ±56%      ~     (p=0.233 n=15+15)
SetOperation/n=256/Go_Union-12                                 1.00ms ±33%    1.24ms ±78%      ~     (p=0.158 n=14+15)
SetOperation/n=256/GEOS_Intersection-12                         229µs ±30%     225µs ±18%      ~     (p=0.870 n=15+15)
SetOperation/n=256/GEOS_Difference-12                           309µs ±27%     289µs ±11%      ~     (p=0.148 n=15+15)
SetOperation/n=256/GEOS_SymmetricDifference-12                  761µs ±38%     728µs ±26%      ~     (p=0.486 n=15+15)
SetOperation/n=256/GEOS_Union-12                                312µs ±13%     312µs ±10%      ~     (p=0.780 n=14+15)
SetOperation/n=512/Go_Intersection-12                          1.75ms ±19%    1.71ms ±18%      ~     (p=0.769 n=14+14)
SetOperation/n=512/Go_Difference-12                            2.16ms ±80%    1.86ms ±38%      ~     (p=0.591 n=15+14)
SetOperation/n=512/Go_SymmetricDifference-12                   2.76ms ±43%    2.67ms ±41%      ~     (p=0.252 n=14+15)
SetOperation/n=512/Go_Union-12                                 1.83ms ±19%    1.82ms ±19%      ~     (p=1.000 n=12+12)
SetOperation/n=512/GEOS_Intersection-12                         401µs ±23%     385µs ±14%      ~     (p=0.451 n=15+14)
SetOperation/n=512/GEOS_Difference-12                           501µs ±29%     484µs ±18%      ~     (p=0.902 n=15+15)
SetOperation/n=512/GEOS_SymmetricDifference-12                 1.26ms ±25%    1.19ms ±16%      ~     (p=0.148 n=15+15)
SetOperation/n=512/GEOS_Union-12                                551µs ±20%     530µs ±14%      ~     (p=0.246 n=14+14)
…