@peterstace 2022-05-13
Ahh, I can see what's happening with CI now. Because of the Go image upgrade for `geos.Dockerfile`, the version of GEOS available with that particular version of Ubuntu is no longer available. I'll see if I can work out a fix.